// src/app/api/models/page-bc-analysis-by-org-hierarchy-response.ts
/* tslint:disable */
/* eslint-disable */
import { BcAnalysisByOrgHierarchyResponse } from './bc-analysis-by-org-hierarchy-response';
import { PageableObject } from './pageable-object';
import { SortObject } from './sort-object';
export interface PageBcAnalysisByOrgHierarchyResponse {
  content?: Array<BcAnalysisByOrgHierarchyResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}